Don't use the register pattern for voting systems
