+++ packages/lesswrong/components/comments/DebateResponse.tsx
@@ -4,7 +4,7 @@ import {useCurrentUser} from '../common/withUser';
 import {DebateResponseWithReplies} from './DebateResponseBlock';
 import classNames from 'classnames';
 import {useVote} from '../votes/withVote';
-import {getVotingSystemByName} from '../../lib/voting/votingSystems';
+import {getVotingSystemByName} from '../../lib/voting/getVotingSystem';
 import type { ContentItemBody } from '../common/ContentItemBody';
 
 const styles = (theme: ThemeType) => ({
