+++ packages/lesswrong/components/votes/withVote.ts
@@ -4,7 +4,8 @@ import { useDialog } from '../common/withDialog';
 import { useMutation, gql } from '@apollo/client';
 import { setVoteClient } from '../../lib/voting/vote';
 import { isAF } from '../../lib/instanceSettings';
-import { VotingSystem, getDefaultVotingSystem } from '../../lib/voting/votingSystems';
+import { getDefaultVotingSystem } from '../../lib/voting/getVotingSystem';
+import { VotingSystem } from '@/lib/voting/votingSystems';
 import * as _ from 'underscore';
 import { VotingProps } from './votingProps';
 import { fragmentTextForQuery } from "../../lib/vulcan-lib/fragments";
